backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
tensorflow==2.15.0
tensorflow-hub==0.15.0
pytesseract==0.3.10
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0
loguru==0.7.2
pytest==7.4.3
python-jose==3.3.0